make recap page work
