+++ src/App.css
@@ -80,21 +80,6 @@
   }
 }
 
-/* Add these animation styles */
-@keyframes negativeScoreShake {
-  0%, 100% { transform: translateX(0); border-color: hsl(var(--destructive)); }
-  25% { transform: translateX(5px); }
-  75% { transform: translateX(-5px); }
-}
-
-.negative-score {
-  animation: negativeScoreShake 0.5s ease-in-out;
-  border-width: 2px;
-  border-color: hsl(var(--destructive));
-}
-
-
-
 @layer base {
   * {
     @apply border-border outline-ring/50;
